Move tier and requirement markers out of fixtures
Add customer-scenario filter and marker
register the markers
default customerscenario
update pytest-polarion-collect and other polarion packages

Remove test injection from test_advanced_search, for meta

The polarion meta was not correctly parsed, and with updates to pytest-polarion-collect and polarion_docstrings, the unmerged docstrings are causing exceptions and failures during collection.

Go to boilerplate for this test module, until the meta processing is resolved.

Update smem to run with python2 explicitly