Apply BZ#1741481 workaround unconditionally.
Allowance of db replication is required for db backup opertaions to work.
This patch is needed as long as we have images with  rewritten pg_hba.conf.
This is done in "loosening" DB permissions step. This step resulted in
the replication being actually disallowed. The vanilla state of the
pg_hba.conf allows replication.