Fix test_provision_stack and test_retire_stack
The stack creation was failing
 * Changed name of the input field from dialog_name to label on 5.11.
 * Need to select correct network for the stack.
 * On 5.11, there is a bug BZ 1754543 making fill method silently fail
   to fill the orchestration template content (it fills only first
   character).  The template then doesn't contain parameter which then
   "surprises" the test as it has no options to fill when ordering the
   stack.
 * Wait for Submit button to get displayed.

The stack was also not properly cleaned-up after I got it created. When
the stack was created using a service, we need to retire the service to
get the stack deleted by CFME.

When we are deleting the stack on EC2, the stack remains in
DELETE_COMPLETE state for a bit, so Stack.exists property needed to be
modified to consider these stacks as non-existent.