Change default provider marker selector to ONE_PER_VERSION
Also update ID's when using the ALL selector to append the provider key (legacy ID)

When encountering multiple providers with the same type+version we were relying on pytest to make our ID's unique, resulting in IDs with an integer directly appended to the ID string.

This resulted in a delta between parametrization against a large provider set and a smaller subset.

I don't think we want ALL to be the default