[1LP][RFR] No more session fixtures (#9639)
there is a plan to implement appliance markers and generate appliances based on those markers.
We should get rid of session scoped fixuters using appliance fixture. This PR is one of steps we have to undertake to achieve above.

Adding hook registering and calling for the appliance stack, in order to call methods on appliances, whenever they're used for tests. Replaces previous session scoped fixtures that were modifying the appliance at the start of the session.