Allow passing generator as the values for fill().
When passing a generator, the generator was depleted in conversion to
a set in first condition. Better approach is to convert the values to
set once and then use it several times instead of doing the conversions several
times.

This commit is also simplifying some unecessarily complicated code in
all_dashboard_widgets().