[1LP][RFR] Fix dashboard tests. (#9684)
* Fix dashboard tests.

Recent changes to widgetastic broke the dashboard tests:

/var/ci/cfme_venv3/lib64/python3.7/site-packages/widgetastic/widget/base.py:1199: AttributeError
AttributeError
b'This is not an instance of widgets. You need to call this object and pass the required parameters of the view.'

* Allow passing generator as the values for fill().

When passing a generator, the generator was depleted in conversion to
a set in first condition. Better approach is to convert the values to
set once and then use it several times instead of doing the conversions several
times.

This commit is also simplifying some unecessarily complicated code in
all_dashboard_widgets().

* Remove the RSS part of the custom_widgets feature for CFME >= 5.11

The RSS was removed in with BZ#1728328