[1LP][RFR] Fix restore tests third part (#9646)
* Make the basebackup command work on 5.11

The parameter -x was dropped, but there is -X fetch instead, which works the same.

* make the repmgr service working on both 5.10 and 5.11

The service name was changed between CFME 5.10 and 5.11.

* Fix test_appliance_console_restore_db_replicated

* Increase timeout for db backup to complete.

* Add with statement to explicitly close the channel

There is a problem with unfinished jobs on PRT. The PR#8655 suffers from
it a lot. This is an attempt to mitigate the issue by explicitly closing
the channel created by SSHClientInteraction.

* Add wait_for to make the ...restore_db_replicated more reliable.

There is a race between the tests and replication manager to sync.
Adding waiting for the providers to appear in the appl2 should help with
this.

* Use less timeout for CFME 5.11

* Increase  timeout for nfs restore

40s seems to be too little

* Fix SCAP appliance_console menu option number.

The number got changed for the SCAP. We didn't really do any SCAP.

* Use interaction.answer in test_appliance_console_db_restore

This does compress the code a bit.

* Change resc to re.escape

* Remove get_replicated_appliances_with_providers fixture

The fixture is redundant. I rather made the tests use the
replicated_appliances_with_providers with an argument that is a list of
either preupdate appliances or temporary (no-preupdate) appliances
depending on which test is requiring them

* Improve test_appliance_console_restore_pg_basebackup_ansible

The were several issues with this test:
 * There was some leftover code doing reboot.
 * Reboot was causing the login page to not fully load on 5.11.
   I think this is some bug that we should investigate, but for now,
   I want the backup/restore test passing.)
 * The runcmd was used to check an aws service was running, but the
   result was not checked.

revert

* Fix test_update_embedded_ansible_webui on 5.11

* Increasing the timeout in configure_appliance_external_join

Fixes test_appliance_console_restore_db_external

* Wait for embedded ansible instead of checking it is running.

This helps with test_appliance_console_restore_pg_basebackup_ansible

* fix test_appliance_console_restore_db_replicated on 5.11

On 5.11, we seem to be getting the "Embedded ansible" as as "bulit-in
provider".

* Fix the ha update test

The test is failing (only) on PRT. Waiting for the failover monitor to
start using monitoring of the log helps.

* Use db_service rather than shell command in test_update_ha

* Changes relted to fixed BZ 1693189

* Define AP_WELCOME_SCREEN_TIMEOUT

* Fix test_update_scap_webui

* Fix waiting for ansible in test_update_embedded_ansible_webui

Use the generalized waiting method.

* Follow options renumberings.

The CFME 5.11.2.1 brough re-numbering of the menu options (BZ 1790995).
This has to be reflected in the tests code.

* Fix HTTP 503 when accessing api of shutoff instance.