fixing test_retry_migration_plan
It seems there is an issue with below Selenium send key"
self.browser.send_keys(Keys.ENTER, self)", So whenever for
the first time send key gets called, it cause the reload of
browser and which landed in all page of the said collection
and throw the exception "NoSuchElementException".

which causes the failure of TC, This PR will works as work around
to resolve the failure.

Signed-off-by: mnadeem92 <mnadeem@redhat.com>