Correct the steps for rebuilding the repo.
The test_rh_updates rebuilds some RPMs to make us able to test CFME should be able to update find to some future version.

The repo with the rpms rebuilidng tool had to be rebuilt (yeah... (: ) because the deps on new CFME 5.11 started being not satisfied. I corrected the steps for rebuilding the repo. This will for sure get handy in future. This is not an ideal system, but at the moment I don't know about how to test updates to some future version nicer way without spending money on time travel machine.

The test_rh_updates started working after the repo has been updated.