Refactor ansible service fixtures
Scope almost all as function
Broadly teardown by deleting all service requests

They are identified only by their description, and the framework is poorly structured in a way that results in exceptions because of this

Services were module scoped, and when ordered all of their resulting requests in various test functions had the same descriptions. Service Request instances in framework need to be refactored to account for the ID that MIQ assigns to the request.

In order to mitigate this circumstance, the fixture were reworked to delete all service requests present via API during teardown at a function scope.